some AI-powered refactoring
