--- src/ui/pages/list/components/types.ts
@@ -1,7 +1,0 @@
-export type ShoppingViewItem = {
-  id: number;
-  name: string;
-  category: string;
-  quantity: string;
-  got: boolean;
-}
